+++ src/bika/lims/content/abstractanalysis.py
@@ -15,7 +15,7 @@
 # this program; if not, write to the Free Software Foundation, Inc., 51
 # Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 #
-# Copyright 2018-2024 by it's authors.
+# Copyright 2018-2025 by it's authors.
 # Some rights reserved, see README and LICENSE.
 
 import cgi
@@ -23,7 +23,6 @@ import copy
 import json
 import math
 from decimal import Decimal
-from six import string_types
 
 from AccessControl import ClassSecurityInfo
 from bika.lims import api
@@ -41,20 +40,21 @@ from bika.lims.config import UDL
 from bika.lims.content.abstractbaseanalysis import AbstractBaseAnalysis
 from bika.lims.content.abstractbaseanalysis import schema
 from bika.lims.interfaces import IDuplicateAnalysis
-from senaite.core.permissions import FieldEditAnalysisResult
-from senaite.core.permissions import ViewResults
 from bika.lims.utils import formatDecimalMark
 from bika.lims.utils.analysis import format_numeric_result
 from bika.lims.utils.analysis import get_significant_digits
 from bika.lims.workflow import getTransitionActor
 from bika.lims.workflow import getTransitionDate
 from DateTime import DateTime
-from senaite.core.browser.fields.datetime import DateTimeField
 from Products.Archetypes.Field import IntegerField
 from Products.Archetypes.Field import StringField
 from Products.Archetypes.references import HoldingReference
 from Products.Archetypes.Schema import Schema
 from Products.CMFCore.permissions import View
+from senaite.core.browser.fields.datetime import DateTimeField
+from senaite.core.permissions import FieldEditAnalysisResult
+from senaite.core.permissions import ViewResults
+from six import string_types
 
 # A link directly to the AnalysisService object used to create the analysis
 AnalysisService = UIDReferenceField(
@@ -111,9 +111,9 @@ Analyst = StringField(
 # The actual uncertainty for this analysis' result, populated from the ranges
 # specified in the analysis service when the result is submitted.
 Uncertainty = StringField(
-    'Uncertainty',
+    "Uncertainty",
     read_permission=View,
-    write_permission="Field: Edit Result",
+    write_permission=FieldEditAnalysisResult,
     precision=10,
 )
 
@@ -581,14 +581,13 @@ class AbstractAnalysis(AbstractBaseAnalysis):
             # check if the dependency is a string result
             str_result = dependency.getStringResult()
             keyword = dependency.getKeyword()
-            if not result:
-                # Dependency without results found
-                if cascade:
-                    # Try to calculate the dependency result
-                    dependency.calculateResult(override, cascade)
-                    result = dependency.getResult()
-                else:
-                    return False
+
+            # Dependency without results found
+            if not result and cascade:
+                # Try to calculate the dependency result
+                dependency.calculateResult(override, cascade)
+                result = dependency.getResult()
+
             if result:
                 try:
                     # we need to quote a string result because of the `eval` below
@@ -611,6 +610,13 @@ class AbstractAnalysis(AbstractBaseAnalysis):
                 # https://docs.python.org/2.7/library/stdtypes.html?highlight=built#string-formatting-operations
                 converter = "s" if str_result else "f"
                 formula = formula.replace("[" + keyword + "]", "%(" + keyword + ")" + converter)
+            else:
+                # flush eventual previously set result
+                if self.getResult():
+                    self.setResult("")
+                    return True
+
+                return False
 
         # convert any remaining placeholders, e.g. from interims etc.
         # NOTE: we assume remaining values are all floatable!